%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  display.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{Display}
\docType{class}
\alias{Display}

 \encoding{latin1}

 \title{The Display class}

 \description{
 Package:   \cr
\bold{Class Display}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\code{Display}\cr

\bold{Directly known subclasses:}\cr
\cr

public static class \bold{Display}\cr
extends \link[R.oo]{Object}\cr



 Serves as a wrapper for a set of graphical functions
 defined as static methods of Display class
 }

 \usage{Display()}

 \arguments{
 \item{...}{Not used.}
 }

 \section{Fields and Methods}{
  \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{\link[IdMappingAnalysis:copy.Display]{copy}} \tab Save current plot to the file.\cr
 \tab \code{\link[IdMappingAnalysis:create.Display]{create}} \tab Open a new display device.\cr
 \tab \code{\link[IdMappingAnalysis:line.loess.Display]{line.loess}} \tab Plot loess transformed data.\cr
 \tab \code{\link[IdMappingAnalysis:line.unsorted.Display]{line.unsorted}} \tab Draw a curve from unsorted points.\cr
 \tab \code{\link[IdMappingAnalysis:progressMsg.Display]{progressMsg}} \tab Display a progress message.\cr
 \tab \code{\link[IdMappingAnalysis:textBoundingBox.Display]{textBoundingBox}} \tab Determine the size of the text bounding box.\cr
 \tab \code{\link[IdMappingAnalysis:zoom.pars.Display]{zoom.pars}} \tab Zoom graphics parameters.\cr
}


 \bold{Methods inherited from Object}:\cr
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clone, detach, equals, extend, finalize, gc, getEnvironment, getFields, getInstantiationTime, getStaticInstance, hasField, hashCode, ll, load, objectSize, print, registerFinalizer, save


 }

 \author{Alex Lisovich, Roger Day}
\keyword{classes}
