\docType{data}
\name{examples}
\alias{examples}
\title{IdMappingAnalysis sample data}
\value{
  A list containing the following sample items:
  \item{mrnaExperimentSet}{ MRNA Experiment sample data
  frame} \item{msmsExperimentSet}{ MSMS Experiment sample
  data frame} \item{outcomeMap}{outcome sample data frame}
  \item{identDfList}{list of ID Map sample data frames
  collected from various services} \item{jointIdMap}{sample
  JointIdMap object for exploring the various DBs
  quantitative identifier mapping performance}
  \item{jointIdMap_corr}{sample JointIdMap object for
  exploring the various DBs correlation based identifier
  mapping performance} \item{uniquePairs}{sample
  UniquePairs object} \item{jointUniquePairs}{sample
  JointUniquePairs object} \item{corrData}{sample CorrData
  object suitable for subsequent fast correlations}
  \item{corr}{sample Corr object encapsulating the
  correlation results} \item{mixture}{sample Mixture object
  encapsulating the mixture modeling results}
  \item{bootstrap}{sample Bootstrap object encapsulating
  the results of bootstrapping on correlations}
}
\description{
  IdMappingAnalysis sample data
}
\author{
  Alex Lisovich, Roger Day
}
\keyword{data}
\keyword{datasets}

