%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  dataFilter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{fisherTransform.DataFilter}
\alias{fisherTransform.DataFilter}
\alias{DataFilter.fisherTransform}
\alias{fisherTransform.DataFilter}
\alias{fisherTransform,DataFilter-method}


 \title{Compute the Fisher transform}
 \description{Compute the Fisher transform.}

 \synopsis{fisherTransform.DataFilter(static, x, ...)}
\usage{DataFilter$fisherTransform(x, ...)}

 \arguments{
 \item{x}{Input \code{\link[base]{numeric}} \code{\link[base]{vector}}.}
 \item{...}{Not used}
 }

 \value{
 Result of the Fisher transform.
 }

 \examples{\dontrun{
 z<-Filter$fisherTransform(x);
 }}

 \seealso{For more information see \code{\link{DataFilter}}.}
 \author{Alex Lisovich, Roger Day}
\keyword{internal}
\keyword{methods}
