\name{Generic R.00 objects}
\alias{Generic R.00 objects}
\title{Generic R.00 objects}

\alias{aligned}
\alias{as.IdMap}
\alias{as.MultiSet}
\alias{as.UniquePairs}
\alias{boxplotdataJitter}
\alias{byColNames}
\alias{byColumn}
\alias{byRow}
\alias{byRowNames}
\alias{clust}
\alias{copy}
\alias{corr.boxplot}
\alias{corr.plot}
\alias{create}
\alias{CsvList.merge}
\alias{diffCounts.plot}
\alias{do.apply}
\alias{do.glm}
\alias{ecdf.plot}
\alias{fisherTransform}
\alias{fisherTransformInverse}
\alias{fisherTransformJacobean}
\alias{getBootstrap}
\alias{getCompoundEvents}
\alias{getCompoundGroups}
\alias{getCorr}
\alias{getCorrData}
\alias{getCorrDataFrame}
\alias{getCounts}
\alias{getData}
\alias{getDiff}
\alias{getExperimentSet}
\alias{getIdMapList}
\alias{getMapNames}
\alias{getMatch}
\alias{getMatchInfo}
\alias{getMixture}
\alias{getSampleNames}
\alias{getStats}
\alias{getUnionIdMap}
\alias{getUniquePairs}
\alias{interactive.corr.boxplot}
\alias{interactive.corr.plot}
\alias{interactive.mixture.boxplot}
\alias{interactive.mixture.plot}
\alias{interactive.plot}
\alias{interleave}
\alias{line.loess}
\alias{line.unsorted}
\alias{minAvgCountConstraint}
\alias{minCountConstraint}
\alias{minCountGroupConstraint}
\alias{mixture.boxplot}
\alias{mixture.plot}
\alias{pack.experiments}
\alias{primaryIDs}
\alias{primaryKey}
\alias{progressMsg}
\alias{removeNASeries}
\alias{secondaryKey}
\alias{subsetCorr}
\alias{subsetData}
\alias{subsetGroups}
\alias{swapKeys}
\alias{textBoundingBox}
\alias{to.base}
\alias{to.binary.logical}
\alias{to.index.expr}
\alias{words}
\alias{zoom.pars}

\description{
R.oo generic S3 methods
}

\keyword{documentation}
\keyword{internal}
