%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  R/jointUniquePairs.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getMapNames.JointUniquePairs}
\alias{getMapNames.JointUniquePairs}
\alias{JointUniquePairs.getMapNames}
\alias{getMapNames,JointUniquePairs-method}

 \title{Get the names of UniquePairs objects encapsulated within the given JointUniquePairs object}
 \description{Get the names of UniquePairs objects encapsulated within the given JointUniquePairs object.}

 \usage{
\method{getMapNames}{JointUniquePairs}(this, ...)
}

 \arguments{\item{...}{Not used}}

 \value{A \code{\link[base]{character}} vector of IdMap names}

 \examples{
 examples$jointUniquePairs$getMapNames();
 }

 \seealso{For more information see \code{\link{JointUniquePairs}}.}
 \author{Alex Lisovich, Roger Day}
\keyword{internal}
\keyword{methods}
