%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  idMapBase.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getName.IdMapBase}
\alias{getName.IdMapBase}
\alias{IdMapBase.getName}
\alias{getName.IdMapBase}
\alias{getName,IdMapBase-method}


 \title{Get the name a given IdMapBase object}
 \description{Get the name a given IdMapBase object.}

 \usage{\method{getName}{IdMapBase}(this, ...)}

 \arguments{\item{...}{Not used}}
 \value{A \code{\link[base]{character}} string representing the name of a given IdMapBase object}

 \seealso{For more information see \code{\link{IdMapBase}}.}
 \author{Alex Lisovich, Roger Day}
\keyword{internal}
\keyword{methods}
