%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  R/corrData.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getSampleNames.CorrData}
\alias{getSampleNames.CorrData}
\alias{CorrData.getSampleNames}
\alias{getSampleNames,CorrData-method}

 \title{Get experiment sample names}

 \description{Get experiment sample names.}

 \usage{
\method{getSampleNames}{CorrData}(this, ...)
}

 \arguments{
 \item{...}{Not used}
 }

 \value{
 \code{\link[base]{character}} vector of sample names encapsulated within the given object.
 }

 \examples{
 examples$corrData$getSampleNames();
 }
 \seealso{For more information see \code{\link{CorrData}}.}
 \author{Alex Lisovich, Roger Day}
\keyword{internal}
\keyword{methods}
