\name{IdMappingAnalysis-package}
\title{IdMappingAnalysis.}
\description{IdMappingAnalysis. Critically comparing identifier maps 
retrieved from bioinformatics annotation resources}
\details{\tabular{ll}{
Package: \tab IdMappingAnalysis\cr
Type: \tab Package\cr
Version: \tab 0.0.5 \cr
Date: \tab 2012-03-07 \cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}}
\alias{IdMappingAnalysis}
\docType{package}
\author{Alex Lisovich, Roger Day}
\keyword{package}
