\docType{package}
\name{IdMappingAnalysis-package}
\alias{IdMappingAnalysis}
\alias{IdMappingAnalysis-package}
\title{IdMappingAnalysis. Critically comparing identifier maps
retrieved from bioinformatics annotation resources}
\description{
  \tabular{ll}{ Package: \tab IdMappingAnalysis\cr Type:
  \tab Package\cr Version: \tab 0.0.5 \cr Date: \tab
  2012-03-07 \cr License: \tab GPL (>= 2)\cr LazyLoad: \tab
  yes\cr }
}
\author{
  Alex Lisovich, Roger Day
}
\keyword{package}

