%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  corrData.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{primaryKey.CorrData}
\alias{primaryKey.CorrData}
\alias{CorrData.primaryKey}
\alias{primaryKey.CorrData}
\alias{primaryKey,CorrData-method}


 \title{Retrieves a primary key for a given CorrData object}
 \description{Retrieves a primary key for a given CorrData object.}

 \usage{\method{primaryKey}{CorrData}(this, ...)}

 \arguments{\item{...}{Not used}}
 \value{A \code{\link[base]{character}} string representing primary key for given CorrData object}

 \examples{
 examples$corrData$primaryKey();
 }

 \seealso{For more information see \code{\link{CorrData}}.}
 \author{Alex Lisovich, Roger Day}
\keyword{internal}
\keyword{methods}
