%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  mixture.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{primaryKey.Mixture}
\alias{primaryKey.Mixture}
\alias{Mixture.primaryKey}
\alias{primaryKey.Mixture}
\alias{primaryKey,Mixture-method}


 \title{Retrieves a primary key for a given Mixture object}

 \description{Retrieves a primary key for a given Mixture object.}

 \usage{\method{primaryKey}{Mixture}(this, ...)}

 \arguments{\item{...}{Not used}}

 \value{A \code{\link[base]{character}} string representing primary key for given Mixture object }

 \examples{
 examples$mixture$primaryKey();
 }

 \seealso{For more information see \code{\link{Mixture}}.}
 \author{Alex Lisovich, Roger Day}
\keyword{internal}
\keyword{methods}
