%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  R/corrData.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{secondaryKey.CorrData}
\alias{secondaryKey.CorrData}
\alias{CorrData.secondaryKey}
\alias{secondaryKey,CorrData-method}

 \title{Retrieves a secondary key for a given CorrData object}
 \description{Retrieves a secondary key for a given CorrData object.}

 \usage{
\method{secondaryKey}{CorrData}(this, ...)
}

 \arguments{\item{...}{Not used}}
 \value{A \code{\link[base]{character}} string representing secondary key for given CorrData object}

 \examples{
 examples$corrData$secondaryKey();
 }

 \seealso{For more information see \code{\link{CorrData}}.}
 \author{Alex Lisovich, Roger Day}
\keyword{internal}
\keyword{methods}
