%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  R/mixture.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{secondaryKey.Mixture}
\alias{secondaryKey.Mixture}
\alias{Mixture.secondaryKey}
\alias{secondaryKey,Mixture-method}

 \title{Retrieves a secondary key for a given Mixture object}
 \description{Retrieves a secondary key for a given Mixture object.}

 \usage{
\method{secondaryKey}{Mixture}(this, ...)
}

 \arguments{\item{...}{Not used}}

 \value{A \code{\link[base]{character}} string representing secondary key for given Mixture object}

 \examples{
 examples$mixture$primaryKey();
 }

 \seealso{For more information see \code{\link{Mixture}}.}
 \author{Alex Lisovich, Roger Day}
\keyword{internal}
\keyword{methods}
