%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  R/subset.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{Subset}
\docType{class}
\alias{Subset}

 \encoding{latin1}

 \title{The Subset class}

 \description{
 Package:  IdMappingAnalysis \cr
\bold{Class Subset}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\code{Subset}\cr

\bold{Directly known subclasses:}\cr
\cr

public static class \bold{Subset}\cr
extends \link[R.oo]{Object}\cr



 Serves as a wrapper for data frame subsetting functions
 defined as static methods of the Subset class.
 }

 \usage{
Subset(...)
}

 \arguments{
 \item{...}{Not used.}
 }

 \section{Fields and Methods}{
  \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{\link[IdMappingAnalysis:byColNames.Subset]{byColNames}} \tab Extract subset of columns from a data frame or a list of data frames.\cr
 \tab \code{\link[IdMappingAnalysis:byColumn.Subset]{byColumn}} \tab # Extract subset of rows from a data frame or a list of data frames
.\cr
 \tab \code{\link[IdMappingAnalysis:byRow.Subset]{byRow}} \tab Extract subset of columns from a data frame or a list of data frames
.\cr
 \tab \code{\link[IdMappingAnalysis:byRowNames.Subset]{byRowNames}} \tab Extract subset of columns from a data frame or a list of data frames.\cr
}


 \bold{Methods inherited from Object}:\cr
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clearLookupCache, clone, detach, equals, extend, finalize, gc, getEnvironment, getFieldModifier, getFieldModifiers, getFields, getInstantiationTime, getStaticInstance, hasField, hashCode, ll, load, objectSize, print, registerFinalizer, save


 }

 \author{Alex Lisovich, Roger Day}
\keyword{classes}
