%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  R/misc.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{Misc}
\docType{class}
\alias{Misc}

 \encoding{latin1}

 \title{The Misc class}

 \description{
 Package:  IdMappingAnalysis \cr
\bold{Class Misc}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\code{Misc}\cr

\bold{Directly known subclasses:}\cr
\cr

public static class \bold{Misc}\cr
extends \link[R.oo]{Object}\cr



 Serves as a wrapper for various miscalleneous functions
 used throughout the package defined as static methods of the Misc class.
 }

 \usage{
Misc()
}

 \arguments{
 }

 \section{Fields and Methods}{
  \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{\link[IdMappingAnalysis:CsvList.merge.Misc]{CsvList.merge}} \tab Pairwise merge of two string vectors.\cr
 \tab \code{\link[IdMappingAnalysis:interleave.Misc]{interleave}} \tab Interleave two matrixes by columns.\cr
 \tab \code{\link[IdMappingAnalysis:to.base.Misc]{to.base}} \tab Convert number to a numeric vector of a given base.\cr
 \tab \code{\link[IdMappingAnalysis:to.binary.logical.Misc]{to.binary.logical}} \tab Convert number to a  vector of logicals.\cr
 \tab \code{\link[IdMappingAnalysis:to.index.expr.Misc]{to.index.expr}} \tab Convert expression into index expression for a given list or data frame object.\cr
 \tab \code{\link[IdMappingAnalysis:words.Misc]{words}} \tab Convert space delimited string to a vector of words.\cr
}


 \bold{Methods inherited from Object}:\cr
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clearLookupCache, clone, detach, equals, extend, finalize, gc, getEnvironment, getFieldModifier, getFieldModifiers, getFields, getInstantiationTime, getStaticInstance, hasField, hashCode, ll, load, objectSize, print, registerFinalizer, save


 }

 \author{Alex Lisovich, Roger Day}
\keyword{classes}
